[Move] 디렉토리 구조 변경에 따른 파일 경로 수정
