+++ src/main/java/com/pinup/controller/TestController.java
@@ -1,4 +1,4 @@
-package com.pinup.domain.test.controller;
+package com.pinup.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
