+++ src/main/java/com/pinup/global/aop/LogAop.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Component;
 public class LogAop {
 
     // controller와 service method만 로그 적용
-    @Around("execution(* com.pinup.domain..controller..*(..)) || execution(* com.pinup.domain..service..*(..))")
+    @Around("execution(* com.pinup.controller..*(..)) || execution(* com.pinup.service..*(..))")
     public Object execute(ProceedingJoinPoint joinPoint) throws Throwable {
         log.info("START: {}", joinPoint.toString());
         try {
