+++ src/main/java/com/pinup/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.pinup.repository;
 
-import com.pinup.domain.member.entity.Member;
+import com.pinup.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
