+++ src/main/java/com/pinup/service/AuthService.java
@@ -1,9 +1,9 @@
 package com.pinup.service;
 
 
-import com.pinup.domain.member.entity.Member;
+import com.pinup.entity.Member;
 
-import com.pinup.dto.request.token.TokenResponse;
+import com.pinup.global.response.TokenResponse;
 import com.pinup.global.enums.LoginType;
 import com.pinup.global.exception.PinUpException;
 import com.pinup.global.jwt.JwtTokenProvider;
